adapting gripper control

+++ scripts/gripper_control_from_topic.py
@@ -31,12 +31,15 @@ class GripperControl:
         # potentially retrieve information from yaml file -> not done in current state
         # self.device_path = rospy.get_param(rospy.get_name() + "/device_path")
 
+        # added some variables to prevent consequtive calling - TODO: maybe improve using mutex!
+        self.gripper_open = None
+        self.currently_opening = False
+        self.currently_closing = False
+
         # first initialize everything
         self.startup_procedure()
 
-
-        self.gripper_open = None
-        self.sub_gripper = rospy.Subscriber('/cartesian_impedance_controller/desired_gripper_state', PointStamped, self.gripper_callback)
+        self.sub_gripper = rospy.Subscriber('/cartesian_impedance_controller/desired_gripper_state', PointStamped, self.gripper_callback, queue_size=1)
 
         # then run online
         rospy.spin()
@@ -44,30 +47,40 @@ class GripperControl:
     def gripper_callback(self, data):
         # data = data.data
         gripper_open = not(data.point.x==1)
-        if self.gripper_open:
-            if not(gripper_open):
-                self.close_gripper()
-        elif not(self.gripper_open):
-            if (gripper_open):
-                self.open_gripper()
+        # if it is switching at the moment - ignore
+        if not(self.currently_closing) and not(self.currently_closing):
+            if self.gripper_open:
+                if not(gripper_open):
+                    self.close_gripper()
+            elif not(self.gripper_open):
+                if (gripper_open):
+                    self.open_gripper()
 
     def startup_procedure(self):
         # open gripper first
         self.open_gripper()
 
     def open_gripper(self):
-        node_process = Popen(shlex.split('rosrun franka_interactive_controllers franka_gripper_run_node 1'))
-        # messagebox.showinfo("Open Gripper", "Gripper Opened")
-        time.sleep(1)  # Sleep for 1 seconds
-        node_process.terminate()
-        self.gripper_open = True
+        if not(self.currently_closing) and not(self.currently_opening):
+            self.currently_opening = True
+            node_process = Popen(shlex.split('rosrun franka_interactive_controllers franka_gripper_run_node 1'))
+            # messagebox.showinfo("Open Gripper", "Gripper Opened")
+            time.sleep(1)  # Sleep for 1 seconds
+            node_process.terminate()
+            time.sleep(1)
+            self.gripper_open = True
+            self.currently_opening = False
 
     def close_gripper(self):
-        node_process = Popen(shlex.split('rosrun franka_interactive_controllers franka_gripper_run_node 0'))
-        # messagebox.showinfo("Close Gripper", "Gripper Closed")
-        time.sleep(1)  # Sleep for 1 seconds
-        node_process.terminate()
-        self.gripper_open = False
+        if not (self.currently_closing) and not (self.currently_opening):
+            self.currently_closing = True
+            node_process = Popen(shlex.split('rosrun franka_interactive_controllers franka_gripper_run_node 0'))
+            # messagebox.showinfo("Close Gripper", "Gripper Closed")
+            time.sleep(1)  # Sleep for 1 seconds
+            node_process.terminate()
+            time.sleep(1)
+            self.gripper_open = False
+            self.currently_closing = False
 
 if __name__ == '__main__':
     try:
